FrontDesk/requirements.txt
livekit==0.12.1
livekit-api==0.6.0
python-dotenv==1.0.1
aiohttp==3.10.5
openai==1.50.2
requests==2.32.3
fastapi==0.115.0
uvicorn==0.30.1
sqlmodel==0.0.22
pydantic==2.9.0
requests==2.32.3
python-dotenv==1.0.1
aiohttp==3.10.5
livekit==0.12.1
livekit-api==0.6.0
streamlit==1.39.0
requests==2.32.3
python-dotenv==1.0.1
pandas==2.2.3